requirements.txt
# === AWS CDK v2 Core Libraries ===
aws-cdk-lib==2.160.0
constructs>=10.0.0

# === Your app dependencies ===
boto3==1.40.68
botocore==1.40.68
jmespath==1.0.1
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.14.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
